third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/source-url-comment-expected.txt

Tests that evals with sourceURL comment are shown in scripts panel.


Running: testSourceURLCommentInInlineScript

Running: testSourceURLCommentInScript
function scriptWithSourceURL() {}
hasSourceURL: true

Running: testPoorSourceURLCommentInScript

Running: testSourceURLComment
function keepAlive() {}
hasSourceURL: true

Running: testSourceURLAndMappingURLComment
function keepAlive() {}
//# sourceMappingURL=sourceMappingURL.map
hasSourceURL: true
sourceMapURL: sourceMappingURL.map

Running: testSourceURLCommentInDynamicScript
function keepAliveInDynamicScript() {}
hasSourceURL: true

Running: testNonRelativeURL
function relativeURLScript() {}
hasSourceURL: true

Running: testMultipleSourceURLComment

//# sourceURL=evalURL2.js
function keepAlive() {}
hasSourceURL: true
